Calculate connection latency and skip too slow connections
Each connections keeps a history of each execute (and executePrepared)
request durations in milliseconds, which allows _getAConnection()
to skip connection if its average latency is too high (configured
with latencyLimit option).

If all connections in the pool have too high latency then a connection
is returned regardless its average latency.

There's a separated Client.getConnectionLatencies() function
which returns a map from server host:port to its average latency,
which can be used to track latencies in the code.

In addition the default consistency level was changed from
ONE to QUORUM in the test suite, because it caused test instability
when using three node cluster against the test suite.